+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### 🛠 Improvements
 - [491](https://github.com/vegaprotocol/vega/issues/491) - Migrate MacOS notification to the UserNotification framework. However, this is still not working. Application is now brought to front when a review or an input from the user is required.
+- [476](https://github.com/vegaprotocol/vega/issues/476) - Import fairground, mainnet1, and testnet2 auto-magically on every startup
 
 ### 🐛 Fixes
 - [473](https://github.com/vegaprotocol/vega/issues/473) - Clean up the application configuration if set with a non-existing network